% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logos_pdf.R
\name{add_logos_pdf}
\alias{add_logos_pdf}
\title{add logos pdf}
\usage{
add_logos_pdf(doc_prefix = NULL)
}
\arguments{
\item{doc_prefix}{Path to pdf_theme_prefix. Default (NULL) will
look up where package is located for default document.}
}
\description{
Set logos as environment variables for fractalSlides design
}
