% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{apply_slide_window}
\alias{apply_slide_window}
\title{Apply a function to each window in a sliding window}
\usage{
apply_slide_window(data, f, key = NULL, w_size = 10, .to = ".out",
  .keep_data = T, ...)
}
\arguments{
\item{data}{A data.frame or list of data points to partition}

\item{f}{function to apply to each window}

\item{key}{Column name to use a key when partitioning,
used only when data is data.frame}

\item{w_size}{Size of the sliding window}

\item{.to}{Name of the column to store the function values in}

\item{.keep_data}{whether to keep the partitioned data (defaults to TRUE)}

\item{...}{Arguments that should be passed to the slide_window function}
}
\value{
A nested data.frame with data partitioned in the data column,
 a `key` column containing windows keys and `w_size` with window size
 and a column named after f or the value specified in .to containing the
 value using the function f
}
\description{
Apply a function to each window in a sliding window
}
