% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{create_windows}
\alias{create_windows}
\title{Create windows of data, using keys in column key}
\usage{
create_windows(data, key = NULL, w_size = 10, ...)
}
\arguments{
\item{data}{A data.frame or list of data points}

\item{key}{Name of column in data (if data is data.frame) that should be used
as key for each window.}

\item{...}{Arguments that should be passed to the create_window function}

\item{fun}{Function that should be applied to each window}
}
\value{
A data.frame with all windows along with corresponding key.
}
\description{
Create windows of data, using keys in column key
}

