% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hop.R
\name{hop}
\alias{hop}
\alias{hop_vec}
\title{Hop}
\usage{
hop(.x, .starts, .stops, .f, ...)

hop_vec(.x, .starts, .stops, .f, ..., .ptype = list())
}
\description{
\code{hop()} jumps along \code{.x} according the the \code{.starts} and \code{.stops}, applying
\code{.f} to each sub-window of \code{.x}.
}
