% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hop2.R, R/phop.R
\name{hop2}
\alias{hop2}
\alias{hop2_vec}
\alias{hop2_dbl}
\alias{hop2_int}
\alias{hop2_lgl}
\alias{hop2_chr}
\alias{hop2_raw}
\alias{hop2_dfr}
\alias{hop2_dfc}
\alias{phop}
\alias{phop_vec}
\alias{phop_dbl}
\alias{phop_int}
\alias{phop_lgl}
\alias{phop_chr}
\alias{phop_raw}
\alias{phop_dfr}
\alias{phop_dfc}
\title{Hop along multiple inputs simultaneously}
\usage{
hop2(.x, .y, .starts, .stops, .f, ...)

hop2_vec(.x, .y, .starts, .stops, .f, ..., .ptype = list())

hop2_dbl(.x, .y, .starts, .stops, .f, ...)

hop2_int(.x, .y, .starts, .stops, .f, ...)

hop2_lgl(.x, .y, .starts, .stops, .f, ...)

hop2_chr(.x, .y, .starts, .stops, .f, ...)

hop2_raw(.x, .y, .starts, .stops, .f, ...)

hop2_dfr(
  .x,
  .y,
  .starts,
  .stops,
  .f,
  ...,
  .names_to = NULL,
  .name_repair = c("unique", "universal", "check_unique")
)

hop2_dfc(
  .x,
  .y,
  .starts,
  .stops,
  .f,
  ...,
  .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)

phop(.l, .starts, .stops, .f, ...)

phop_vec(.l, .starts, .stops, .f, ..., .ptype = list())

phop_dbl(.l, .starts, .stops, .f, ...)

phop_int(.l, .starts, .stops, .f, ...)

phop_lgl(.l, .starts, .stops, .f, ...)

phop_chr(.l, .starts, .stops, .f, ...)

phop_raw(.l, .starts, .stops, .f, ...)

phop_dfr(
  .l,
  .starts,
  .stops,
  .f,
  ...,
  .names_to = NULL,
  .name_repair = c("unique", "universal", "check_unique")
)

phop_dfc(
  .l,
  .starts,
  .stops,
  .f,
  ...,
  .size = NULL,
  .name_repair = c("unique", "universal", "check_unique", "minimal")
)
}
\arguments{
\item{.x, .y}{\verb{[vector]}

Vectors to iterate over. Vectors of size 1 will be recycled.}

\item{.starts}{\verb{[vector]}

Vectors of boundary locations that make up the windows to bucket \code{.x} with.
Both \code{.starts} and \code{.stops} will be recycled to their common size, and
that common size will be the size of the result. Both vectors should be
integer locations along \code{.x}.}

\item{.stops}{\verb{[vector]}

Vectors of boundary locations that make up the windows to bucket \code{.x} with.
Both \code{.starts} and \code{.stops} will be recycled to their common size, and
that common size will be the size of the result. Both vectors should be
integer locations along \code{.x}.}

\item{.f}{\verb{[function / formula]}

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function with up
to two arguments: \code{.x} (single argument) or \code{.x} and \code{.y} (two arguments).
The \code{.} placeholder can be used instead of \code{.x}. This allows you to
create very compact anonymous functions with up to two inputs.}

\item{...}{Additional arguments passed on to the mapped function.}

\item{.ptype}{\verb{[vector]}

The prototype corresponding to the type of the output. Defaults to
a \code{list()}.}

\item{.names_to}{Optionally, the name of a column where the names
of \code{...} arguments are copied. These names are useful to identify
which row comes from which input. If supplied and \code{...} is not named,
an integer column is used to identify the rows.}

\item{.name_repair}{One of \code{"unique"}, \code{"universal"}, or
\code{"check_unique"}. See \code{\link[vctrs:vec_as_names]{vec_as_names()}} for the meaning of these
options.

With \code{vec_rbind()}, the repair function is applied to all inputs
separately. This is because \code{vec_rbind()} needs to align their
columns before binding the rows, and thus needs all inputs to
have unique names. On the other hand, \code{vec_cbind()} applies the
repair function after all inputs have been concatenated together
in a final data frame. Hence \code{vec_cbind()} allows the more
permissive minimal names repair.}

\item{.size}{If, \code{NULL}, the default, will determine the number of
rows in \code{vec_cbind()} output by using the standard recycling rules.

Alternatively, specify the desired number of rows, and any inputs
of length 1 will be recycled appropriately.}

\item{.l}{\verb{[list]}

A list of vectors. The length of \code{.l} determines the
number of arguments that \code{.f} will be called with. If \code{.l} has names,
they will be used as named arguments to \code{.f}. Elements of \code{.l} with size
1 will be recycled.}
}
\description{
\code{hop2()} and \code{phop()} represent the combination
of \code{\link[=slide2]{slide2()}} and \code{\link[=pslide]{pslide()}} with \code{\link[=hop]{hop()}}, allowing you to iterate
over multiple vectors at once, hopping along them using boundaries defined
by \code{.starts} and \code{.stops}.
}
\section{Invariants}{


\subsection{\code{hop2()}}{
\itemize{
\item \code{vec_size(hop2(.x, .y, .starts, .stops)) == vec_size_common(.starts, .stops)}
\item \code{vec_ptype(hop2(.x, .y, .starts, .stops)) == list()}
}

}

\subsection{\code{hop2_vec()} and \verb{hop2_*()} variants}{
\itemize{
\item \code{vec_size(hop2_vec(.x, .y, .starts, .stops)) == vec_size_common(.starts, .stops)}
\item \code{vec_size(hop2_vec(.x, .y, .starts, .stops)[[1]]) == 1L}
\item \code{vec_ptype(hop2_vec(.x, .y, .starts, .stops, .ptype = ptype)) == ptype}
}

}

\subsection{\code{phop()}}{
\itemize{
\item \code{vec_size(phop(.l, .starts, .stops)) == vec_size_common(.starts, .stops)}
\item \code{vec_ptype(phop(.l, .starts, .stops)) == list()}
}

}

\subsection{\code{phop_vec()} and \verb{phop_*()} variants}{
\itemize{
\item \code{vec_size(phop_vec(.l, .starts, .stops)) == vec_size_common(.starts, .stops)}
\item \code{vec_size(phop_vec(.l, .starts, .stops)[[1]]) == 1L}
\item \code{vec_ptype(phop_vec(.l, .starts, .stops, .ptype = ptype)) == ptype}
}

}
}

\examples{

hop2(1:2, 3:4, .starts = 1, .stops = c(2, 1), ~c(x = .x, y = .y))

phop(
 list(1, 2:4, 5:7),
 .starts = c(0, 1),
 .stops  = c(2, 4),
 ~c(x = ..1, y = ..2, z = ..3)
)

}
\seealso{
\code{\link[=hop]{hop()}}, \code{\link[=hop_index]{hop_index()}}, \code{\link[=slide2]{slide2()}}
}
