\name{LRTPermTest}
\Rdversion{1.1}
\alias{LRTPermTest}

\title{
Perform permutation test on the Likelihood Ratio Statistics 
}
\description{
Perform permutation test on the Likelihood Ratio Statistics for each spot from the gel data
}
\usage{
LRTPermTest(LRTStat, data, noP = 1000 , th=NA, tMax=log2(100))
}

\arguments{
  \item{LRTStat}{
	Likelihood Ratio Statistics generated by gelLRT method
}
  \item{data}{
	Numeric n by g matrix records log intensities from gel data. Where n is the number of spots and g is the number of gels.
}
  \item{noP}{
	Number of permutation used for the permutation test. Default 1000.
	}
  \item{th}{
	Numeric values for the limits of detection. Default value is the minimum value from gel data.
}
\item{tMax}{
        Theoretical maximum, default log2(100)
}
}

%\details{
% ~~further details
%}
\value{
    Return a vector of length n which contain the p-value based on the permutation test
}
%%\references{
%% ~put references to the literature/web site here ~
%%}
\author{
S. Wu
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link[slider:simGel]{simGel}}, \code{\link[stats:Chisquare]{pchisq}}, \code{\link[stats:p.adjust]{p.adjust}}
}
\examples{
data(simGel)
result<- gelLRT(simGel$data,simGel$group)
\dontrun{ LRT.pvalue<- LRTPermTest(result, simGel$data, noP=100) }

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }

