% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{nest_windows}
\alias{nest_windows}
\title{Creates sliding windows from the data and returns them as a nested data.frame}
\usage{
nest_windows(data, key = NULL, w_size = 10, ...)
}
\arguments{
\item{data}{A vector, list or data.frame}

\item{key}{If data is data.frame,
provide a column name that serves as key for each window}

\item{w_size}{The size of the windows}
}
\value{
A nested data.frame with two columns: "key" and "data"
}
\description{
Basically a wrapper around create_windows and nest
}
\examples{
df.test <- data.frame( x=1:100, y=1:100*2 )
nest_windows(df.test, key="x")
}

