% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition_window}
\alias{partition_window}
\title{Create a window of certain size at each index of the data.}
\usage{
partition_window(idx, w_size, keys, data = NULL, partial = T,
  align = "center", ...)
}
\arguments{
\item{idx}{The index of the window key.}

\item{w_size}{Size of window in points.}

\item{keys}{The values used to create and identify the windows.}

\item{data}{The data that will be filtered and divided into windows.}

\item{partial}{Whether a partial window is allowed.
At the begin and end of a dataset, some windows will be smaller than w_size.
These windows will be removed is partial is FALSE.}

\item{align}{Alignment of the window, relative to idx.
"center" means: idx is in the middle between lower and upper bound.
"left" means: idx is on the left, so idx == lower bound.
"right" means: idx is on the right, so idx == upper bound.}
}
\value{
The subset of keys (or data) that are in the window
}
\description{
Create a window of certain size at each index of the data.
}

