% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{resolve.quosure}
\alias{resolve.quosure}
\title{Resolves quosures so that character input given to a quoted argument is resolved to symbol.}
\usage{
resolve.quosure(x)
}
\arguments{
\item{x}{A quoted variable that needs to be resolved}
}
\value{
The resolved quosure
}
\description{
Adapted from https://stackoverflow.com/q/43700765
}
