% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide_window}
\alias{slide_window}
\title{Partition data using a sliding window}
\usage{
slide_window(data, key = NULL, w_size = 10, ...)
}
\arguments{
\item{data}{A data.frame or list of data points}

\item{key}{Name of column in data (if data is data.frame) that should be used
as key for each window.}

\item{w_size}{The size of the window to create}

\item{...}{Arguments that should be passed to the create_window function}
}
\value{
A nested data.frame with all windows along with corresponding key.
}
\description{
Partition data using a sliding window
}
