% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide_windows}
\alias{slide_windows}
\title{Partition data using sliding windows of different sizes}
\usage{
slide_windows(data, key = NULL, w_sizes = 10, ...)
}
\arguments{
\item{data}{A data.frame or list of data points to partition}

\item{key}{Column name to use a key when partitioning,
used only when data is data.frame}

\item{w_sizes}{Sizes of the windows to create}

\item{...}{Arguments that should be passed to the slide_window function}
}
\value{
A nested data.frame with all windows of all sizes,
 with `key` column containing windows keys and `w_size` with window size.
}
\description{
Partition data using sliding windows of different sizes
}
