% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration.R
\name{slider_iteration}
\alias{slider_iteration}
\title{Context dependent information}
\usage{
slider_iteration()
}
\description{
\code{slider_iteration()} returns the current iteration number. It is only usable
from within a \verb{slide_*()} function, and will error otherwise.
}
\examples{
# Without `slider_iteration()`, there is no way to know what iteration
# you are on in complex slides
slide(1:3, ~list(slider_iteration(), .x), .before = 1, .after = 1)

# It respects `.step` too
slide_int(6:10, ~slider_iteration(), .step = 2)
}
