% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbounded.R
\name{unbounded}
\alias{unbounded}
\title{Unbounded}
\usage{
unbounded()
}
\description{
\code{unbounded()} can be used as the \code{.before} or \code{.after} argument to \code{\link[=slide]{slide()}}.
It is generally used to create a \emph{cumulative} sliding window. For example,
when \code{.before = unbounded()} is set, the start of sliding window will
always be pinned to the first position in \code{.x}, while the end of the sliding
window continues to shift forward along \code{.x}.
}
\examples{

slide(1:5, ~.x, .before = unbounded())

}
\seealso{
\code{\link[=slide]{slide()}}
}
