% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-package.R
\docType{package}
\name{slide-package}
\alias{slide-package}
\alias{_PACKAGE}
\title{slide: Sliding Window Functions}
\description{
Provides type stable rolling and cumulative window functions 
    over any R data type.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/slide}
  \item Report bugs at \url{https://github.com/DavisVaughan/slide/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
