% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.R
\name{slide}
\alias{slide}
\alias{slide_dbl}
\alias{slide_dfr}
\title{Slide}
\usage{
slide(.x, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .partial = FALSE, .dir = "forward")

slide_dbl(.x, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .partial = FALSE, .dir = "forward")

slide_dfr(.x, .f, ..., .before = 0L, .after = 0L, .step = 1L,
  .offset = NULL, .partial = FALSE, .dir = "forward",
  .names_to = NULL, .name_repair = c("unique", "universal",
  "check_unique"))
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function or formula.}

\item{...}{Additional arguments passed on to the mapped function.}

\item{.step}{A single positive integer. The size of the step between each
function evaluation.}

\item{.partial}{Should partial results be computed?
If \code{TRUE} and \code{.dir == "forward"}, this may result in partial results at the
\emph{end} of the vector. If \code{.dir == "backwards"}, this may result in partial
results at the \emph{start} of the vector.}

\item{.dir}{The direction to slide.

One of:
\itemize{
\item \code{"forward"}
\item \code{"backward"}
}}

\item{.size}{A single positive integer. The size of the sliding window.}

\item{.align}{The alignment to begin the placement of results in the output
vector.

One of:
\itemize{
\item \code{"right"}
\item \code{"left"}
\item \code{"center"}
\item \code{"center-left"}
\item \code{"center-right"}
}

When \code{.size} is even, \code{"center"} is assumed to be
equivalent to \code{"center-left"}.}
}
\description{
Slide
}
