% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide.r
\name{slide}
\alias{slide}
\title{slide}
\usage{
slide(
  trainfile,
  testfile,
  logfile = "/tmp/log.txt",
  batchsize = 1000,
  rehash = 1000,
  rebuild = 1000,
  input_dim = 784,
  Lr = 1e-04,
  epochs = 5,
  stepsize = 20,
  layers = 3
)
}
\arguments{
\item{trainfile, testfile}{TODO}

\item{logfile}{TODO}

\item{batchsize}{TODO}

\item{rehash}{TODO}

\item{rebuild}{TODO}

\item{input_dim}{TODO}

\item{Lr}{TODO}

\item{epochs}{TODO}

\item{stepsize}{TODO}

\item{layers}{TODO}
}
\description{
TODO
}
