% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-between2.R
\name{slide_between2}
\alias{slide_between2}
\alias{slide_between2_vec}
\title{Slide over multiple inputs between boundaries}
\usage{
slide_between2(.x, .y, .i, .starts, .stops, .f, ...)

slide_between2_vec(.x, .y, .i, .starts, .stops, .f, ..., .ptype = list())
}
\description{
Slide over multiple inputs between boundaries
}
