% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider-package.R
\docType{package}
\name{slider-package}
\alias{slider-package}
\alias{_PACKAGE}
\title{slider: Sliding Window Functions}
\description{
Provides type-stable rolling window functions over any R data type. Cumulative and expanding windows are also supported. For more advanced usage, an index can be used as a secondary vector that defines how sliding windows are to be created.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/slider}
  \item \url{https://slider.r-lib.org}
  \item Report bugs at \url{https://github.com/r-lib/slider/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
