% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider-package.r
\docType{package}
\name{slider-package}
\alias{slider-package}
\title{slider-package}
\description{
R bindings for the the Sub-LInear Deep learning Engine (SLIDE) library.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
