% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurrr-package.R
\docType{package}
\name{slurrr-package}
\alias{slurrr}
\alias{slurrr-package}
\title{slurrr: Sliding Window Functions}
\description{
Provides type stable rolling and cumulative window functions 
    over any R data type.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/slurrr}
  \item Report bugs at \url{https://github.com/DavisVaughan/slurrr/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
